+++ lib/benchee_dsl/benchmark.ex
@@ -70,17 +70,8 @@ defmodule BencheeDsl.Benchmark do
 
       @behaviour BencheeDsl.Benchmark
 
-      @before_compile BencheeDsl.Benchmark
-
       Module.register_attribute(__MODULE__, :title, persist: true)
       Module.register_attribute(__MODULE__, :description, persist: true)
-
-      Module.register_attribute(__MODULE__, :tag, accumulate: true)
-
-      Module.register_attribute(__MODULE__, :after_scenario, persist: true)
-      Module.register_attribute(__MODULE__, :after_each, persist: true)
-      Module.register_attribute(__MODULE__, :before_scenario, persist: true)
-      Module.register_attribute(__MODULE__, :before_each, persist: true)
 
       Module.register_attribute(__MODULE__, :__dir__, persist: true)
       Module.put_attribute(__MODULE__, :__dir__, __DIR__)
@@ -88,18 +79,18 @@ defmodule BencheeDsl.Benchmark do
       Module.register_attribute(__MODULE__, :__file__, persist: true)
       Module.put_attribute(__MODULE__, :__file__, __ENV__.file)
 
+      Module.register_attribute(__MODULE__, :tag, accumulate: true)
+
+      Module.register_attribute(__MODULE__, :after_scenario, accumulate: true)
+      Module.register_attribute(__MODULE__, :after_each, accumulate: true)
+      Module.register_attribute(__MODULE__, :before_scenario, accumulate: true)
+      Module.register_attribute(__MODULE__, :before_each, accumulate: true)
+
       @impl BencheeDsl.Benchmark
       @spec run(keyword()) :: :ok
       def run(config \\ []) do
         Server.run(config, %{include: __MODULE__, run: :iex})
       end
-    end
-  end
-
-  defmacro __before_compile__(_env) do
-    quote do
-      @doc false
-      def hook(_name), do: nil
     end
   end
 
@@ -119,26 +110,6 @@ defmodule BencheeDsl.Benchmark do
   defmacro on_exit(fun) do
     quote do
       Server.register(:on_exit, __MODULE__, unquote(fun))
-    end
-  end
-
-  @doc """
-  Overwrites the job map. This function can be used to set up jobs dynamically.
-  """
-  defmacro jobs(do: body) do
-    quote do
-      @doc false
-      def jobs, do: unquote(body)
-    end
-  end
-
-  @doc """
-  Updates the job map. This function can be used to set up jobs dynamically.
-  """
-  defmacro jobs(var, do: body) do
-    quote do
-      @doc false
-      def jobs(unquote(var)), do: unquote(body)
     end
   end
 
@@ -197,58 +168,70 @@ defmodule BencheeDsl.Benchmark do
     quote_job(fun_name, var, body)
   end
 
+  @doc """
+  Takes a module and generates jobs for each publich function.
+  """
+  defmacro jobs(module) do
+    quote bind_quoted: [module: module] do
+      tags = Module.delete_attribute(__MODULE__, :tag)
+
+      Enum.each(module.__info__(:functions), fn {name, arity} ->
+        Server.register(:job, __MODULE__, name,
+          tags: tags,
+          fun: fn input -> apply(module, name, input) end
+        )
+      end)
+    end
+  end
+
   defp quote_job(fun_name, var, body) do
-    quote do
-      unquote(register_job(fun_name))
-
-      @doc false
-      def job(unquote(fun_name)) do
-        fn unquote(var) -> unquote(body) end
-      end
-    end
-  end
-
-  defp quote_job(fun_name, body) do
-    quote do
-      unquote(register_job(fun_name))
-
-      @doc false
-      def job(unquote(fun_name)) do
-        fn -> unquote(body) end
-      end
-    end
-  end
-
-  defp quote_job_apply(fun_name, body, 0) do
-    quote do
-      unquote(register_job(fun_name))
-
-      @doc false
-      def job(unquote(fun_name)) do
-        fn -> apply(unquote(body), []) end
-      end
-    end
-  end
-
-  defp quote_job_apply(fun_name, body, _arity) do
-    quote do
-      unquote(register_job(fun_name))
-
-      @doc false
-      def job(unquote(fun_name)) do
-        fn input -> apply(unquote(body), input) end
-      end
-    end
-  end
-
-  defp register_job(fun_name) do
     quote do
       Server.register(:job, __MODULE__, unquote(fun_name),
         after_each: Module.delete_attribute(__MODULE__, :after_each),
         after_scenario: Module.delete_attribute(__MODULE__, :after_scenario),
         before_each: Module.delete_attribute(__MODULE__, :before_each),
         before_scenario: Module.delete_attribute(__MODULE__, :before_scenario),
-        tags: Module.delete_attribute(__MODULE__, :tag)
+        tags: Module.delete_attribute(__MODULE__, :tag),
+        fun: fn unquote(var) -> unquote(body) end
+      )
+    end
+  end
+
+  defp quote_job(fun_name, body) do
+    quote do
+      Server.register(:job, __MODULE__, unquote(fun_name),
+        after_each: Module.delete_attribute(__MODULE__, :after_each),
+        after_scenario: Module.delete_attribute(__MODULE__, :after_scenario),
+        before_each: Module.delete_attribute(__MODULE__, :before_each),
+        before_scenario: Module.delete_attribute(__MODULE__, :before_scenario),
+        tags: Module.delete_attribute(__MODULE__, :tag),
+        fun: fn -> unquote(body) end
+      )
+    end
+  end
+
+  defp quote_job_apply(fun_name, body, 0) do
+    quote do
+      Server.register(:job, __MODULE__, unquote(fun_name),
+        after_each: Module.delete_attribute(__MODULE__, :after_each),
+        after_scenario: Module.delete_attribute(__MODULE__, :after_scenario),
+        before_each: Module.delete_attribute(__MODULE__, :before_each),
+        before_scenario: Module.delete_attribute(__MODULE__, :before_scenario),
+        tags: Module.delete_attribute(__MODULE__, :tag),
+        fun: fn -> apply(unquote(body), []) end
+      )
+    end
+  end
+
+  defp quote_job_apply(fun_name, body, _arity) do
+    quote do
+      Server.register(:job, __MODULE__, unquote(fun_name),
+        after_each: Module.delete_attribute(__MODULE__, :after_each),
+        after_scenario: Module.delete_attribute(__MODULE__, :after_scenario),
+        before_each: Module.delete_attribute(__MODULE__, :before_each),
+        before_scenario: Module.delete_attribute(__MODULE__, :before_scenario),
+        tags: Module.delete_attribute(__MODULE__, :tag),
+        fun: fn input -> apply(unquote(body), input) end
       )
     end
   end
@@ -267,13 +250,10 @@ defmodule BencheeDsl.Benchmark do
   """
   defmacro before_scenario(do: body) do
     quote do
-      @doc false
-      def hook(:before_scenario) do
-        fn inputs ->
-          unquote(body)
-          inputs
-        end
-      end
+      Server.register(:before_scenario, __MODULE__, fn inputs ->
+        unquote(body)
+        inputs
+      end)
     end
   end
 
@@ -282,8 +262,7 @@ defmodule BencheeDsl.Benchmark do
   """
   defmacro before_scenario(var, do: body) do
     quote do
-      @doc false
-      def hook(:before_scenario), do: fn unquote(var) -> unquote(body) end
+      Server.register(:before_scenario, __MODULE__, fn unquote(var) -> unquote(body) end)
     end
   end
 
@@ -292,13 +271,10 @@ defmodule BencheeDsl.Benchmark do
   """
   defmacro after_scenario(do: body) do
     quote do
-      @doc false
-      def hook(:after_scenario) do
-        fn inputs ->
-          unquote(body)
-          inputs
-        end
-      end
+      Server.register(:after_scenario, __MODULE__, fn inputs ->
+        unquote(body)
+        inputs
+      end)
     end
   end
 
@@ -307,8 +283,7 @@ defmodule BencheeDsl.Benchmark do
   """
   defmacro after_scenario(var, do: body) do
     quote do
-      @doc false
-      def hook(:after_scenario), do: fn unquote(var) -> unquote(body) end
+      Server.register(:after_scenario, __MODULE__, fn unquote(var) -> unquote(body) end)
     end
   end
 
@@ -317,13 +292,10 @@ defmodule BencheeDsl.Benchmark do
   """
   defmacro before_each(do: body) do
     quote do
-      @doc false
-      def hook(:before_each) do
-        fn inputs ->
-          unquote(body)
-          inputs
-        end
-      end
+      Server.register(:before_each, __MODULE__, fn inputs ->
+        unquote(body)
+        inputs
+      end)
     end
   end
 
@@ -332,8 +304,7 @@ defmodule BencheeDsl.Benchmark do
   """
   defmacro before_each(var, do: body) do
     quote do
-      @doc false
-      def hook(:before_each), do: fn unquote(var) -> unquote(body) end
+      Server.register(:before_each, __MODULE__, fn unquote(var) -> unquote(body) end)
     end
   end
 
@@ -342,13 +313,7 @@ defmodule BencheeDsl.Benchmark do
   """
   defmacro after_each(do: body) do
     quote do
-      @doc false
-      def hook(:after_each) do
-        fn inputs ->
-          unquote(body)
-          inputs
-        end
-      end
+      Server.register(:after_each, __MODULE__, fn _result -> unquote(body) end)
     end
   end
 
@@ -357,8 +322,7 @@ defmodule BencheeDsl.Benchmark do
   """
   defmacro after_each(var, do: body) do
     quote do
-      @doc false
-      def hook(:after_each), do: fn unquote(var) -> unquote(body) end
+      Server.register(:after_each, __MODULE__, fn unquote(var) -> unquote(body) end)
     end
   end
 end