+++ test/fixtures/attr_bench.exs
@@ -4,15 +4,14 @@ defmodule AttrBench do
   @title "title"
   @description "description"
 
-  @list Enum.to_list(1..10_000)
-
-  def map_fun(i), do: [i, i * i]
+  list = Enum.to_list(1..10_000)
+  map_fun = fn i -> [i, i * i] end
 
   job flat_map do
-    Enum.flat_map(@list, &map_fun/1)
+    Enum.flat_map(list, map_fun)
   end
 
   job "map.flatten" do
-    @list |> Enum.map(&map_fun/1) |> List.flatten()
+    list |> Enum.map(map_fun) |> List.flatten()
   end
 end