+++ test/test_helper.exs
@@ -7,4 +7,4 @@ Application.put_env(:benchee_dsl, :benchee, BencheeDsl.BencheeMock)
 benchee_run = if System.get_env("CI"), do: true, else: false
 Application.put_env(:benchee_dsl, :benchee_run, benchee_run)
 
-ExUnit.start(capture_log: true)
+ExUnit.start()